Use case structure reformatting
