+++ app/src/main/java/com/example/blanball/utils/di/DomainModule.kt
@@ -1,6 +1,8 @@
 package com.example.blanball.utils.di
 
+import com.example.domain.usecases.implementations.RegistrationUseCaseImpl
 import com.example.domain.usecases.implementations.ResetPasswordUseCaseImpl
+import com.example.domain.usecases.interfaces.RegistrationUseCase
 import com.example.domain.usecases.interfaces.ResetPasswordUseCase
 import dagger.Binds
 import dagger.Module
@@ -12,5 +14,8 @@ import dagger.hilt.components.SingletonComponent
 interface DomainModule {
 @Binds
 fun bindResetPasswordUseCase(resetPasswordUseCase: ResetPasswordUseCaseImpl) : ResetPasswordUseCase
+
+@Binds
+fun bindRegistrationUseCase(registrationUseCase: RegistrationUseCaseImpl) : RegistrationUseCase
 
 }
